requirements.txt
torch>=2.2.0
torchvision>=0.15.2
scikit-learn==1.2.2
numpy==1.23.5
pandas==2.0.1
matplotlib==3.7.1
seaborn==0.12.2
opencv-python==4.7.0.72
tqdm==4.65.0
pyyaml==6.0
pillow==9.5.0
einops==0.6.1
timm==0.9.2